Implement user/movie services

+++ src/main/java/com/example/demo/services/MovieService.java
@@ -4,6 +4,6 @@ import com.example.demo.models.MoviesModel;
 
 public interface MovieService {
     public void addMovie(MoviesModel moviesModel);
-    public void deleteMovie(int id);
+    public void deleteMovie(Long id);
     public void updateMovie(int id);
 }
